third_party/blink/web_tests/fast/scrolling/scrollbars/mouse-autoscrolling-on-scrollbar.html

<!DOCTYPE html>
<title>Tests mouse autoscroll interactions on a non-custom composited div scrollbar.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/scrollbar-util.js"></script>
<style>
.appearance {
  width: 100px;
  height: 100px;
  overflow: scroll;
  border: 1px solid black;
}
.location {
  position: absolute;
  top: 100px;
  left: 100px;
}
.fast {
  will-change: transform;
}
.space {
  height: 2000px;
  width: 2000px;
}

</style>

<!-- Composited non-custom fast scroller -->
<div id="scroller" class="appearance location fast">
  <div class="space"></div>
</div>

<script>
  // Turn off animated scrolling. The "conditionHolds" API expects the scrollTop to
  // *not* change for 10 frames. This will fail since the last GSU would still be
  // animating if animated scrolling were on.
  if (window.internals)
    internals.settings.setScrollAnimatorEnabled(false);

  const scroller = document.getElementById("scroller");
  const scrollerRect = scroller.getBoundingClientRect();

  const TRACK_WIDTH = calculateScrollbarThickness();
  const BUTTON_WIDTH = TRACK_WIDTH;
  const SCROLL_CORNER = TRACK_WIDTH;
  const SCROLL_DELTA = 400;
  const MAX_SCROLLER_OFFSET = 1915;
  const PRESS_DURATION = 1000;
  const down_arrow_x = scrollerRect.right - BUTTON_WIDTH / 2;
  const down_arrow_y = scrollerRect.bottom - SCROLL_CORNER - BUTTON_WIDTH / 2;

  promise_test (async () => {
    await waitForCompositorCommit();
    scroller.scrollTop = 0;

    // TODO(arakeri): Split the mousePressOn API calls to mouseDownAt, waitFor
    // and mouseUpAt once crbug.com/979408 is fixed.
    await mousePressOn(down_arrow_x, down_arrow_y, PRESS_DURATION);
    const err = `Autoscroll down failed (scroller.scrollTop = ${scroller.scrollTop})`;

    // Verify that autoscroll happened.
    assert_greater_than(scroller.scrollTop, SCROLL_DELTA, err);

    // Since autoscroll for arrows happens at 800 px per second, verify that the
    // scrollTop has not reached the end.
    assert_less_than(scroller.scrollTop, MAX_SCROLLER_OFFSET, "Reached scroller end.");

    await waitForCompositorCommit();
    const current_offset = scroller.scrollTop;
    await conditionHolds(() => { return parseInt(scroller.scrollTop) == parseInt(current_offset); },
    `scroller.scrollTop = ${scroller.scrollTop} current_offset = ${current_offset}`);
  },"Test autoscroll down and autoscroll stop.");
</script>
